+++ DebugFileTools/SymbolsRuntime/symbol_file_header.h
@@ -3,9 +3,9 @@
 #pragma pack(push, 1)
 struct s_symbol_file_header
 {
-	unsigned long long public_symbols_count : 24;
-	unsigned long long static_symbols_count : 24;
-	unsigned long long sections_count : 16;
+	uint64_t public_symbols_count : 24;
+	uint64_t static_symbols_count : 24;
+	uint64_t sections_count : 16;
 	union
 	{
 		uintptr_t public_symbols_offset;
@@ -32,14 +32,14 @@ struct s_symbol_file_header
 		uintptr_t timestamp_string_offset;
 		const char* timestamp_string;
 	};
-	unsigned long entry_point_rva;
+	uint32_t entry_point_rva;
 	union
 	{
 		unsigned short entry_point_section_index;
 		struct
 		{
-			unsigned long long : 16;
-			unsigned long long preferred_load_address : 48;
+			uint64_t : 16;
+			uint64_t preferred_load_address : 48;
 		};
 	};
 };
